Fix LOD render distances being overwritten by graphics settings change
Hook the Level Of Detail setting index change event, and refresh ShaderAsset render distances.

To facilitate that a new method is exposed on ShaderAsset.

Also, because 'container buildings' (those that contain a prop that uses the shader) also have their prop render distance overwritten, those are now included in the assets list.

Renamed `allShaders` list to `assets` to be more correct; it's not a list of shaders, it's a list of assets that use the shader (or contain something that does).

Also managed to make the asset scanner 2x faster when scanning `PropInfo` and `BuildingInfo` by using `for` loop approach and by filtering out vanilla items. Before: 109ms startup, Now: 48ms startup :)